+++ cfme/providers/hawkular.py
@@ -1,13 +1,13 @@
 import re
 
 from cfme.common import TopologyMixin, TimelinesMixin
-from . import MiddlewareProvider
+from middleware import MiddlewareProvider
 from mgmtsystem.hawkular import Hawkular
 from utils.appliance import Navigatable
 from utils.db import cfmedb
 from utils.varmeth import variable
-from . import _get_providers_page, _db_select_query
-from .. import download, MiddlewareBase, auth_btn, mon_btn
+from middleware import _get_providers_page, _db_select_query
+from cfme.middleware import download, MiddlewareBase, auth_btn, mon_btn
 from utils.appliance.implementations.ui import navigate_to
