+++ cfme/providers/openshift.py
@@ -1,4 +1,4 @@
-from . import ContainersProvider
+from containers import ContainersProvider
 from utils.varmeth import variable
 from mgmtsystem.openshift import Openshift
 from os import path
